LayoutTests/fast/dom/MicroData/propertiescollection-on-changing-id-attr-expected.txt

Test to verify HTMLPropertiesCollection and PropertyNodeList interfaces behavior on changing id attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Created element of type: div
PASS item.properties.length is 0
PASS item.properties.item(0) is null
PASS item.properties[0] is undefined.
PASS item.properties.namedItem('foo').length is 0
PASS item.properties.namedItem('foo').item(0) is null
PASS item.properties.namedItem('foo')[0] is undefined.
PASS item.properties.namedItem('bar').length is 0
PASS item.properties.namedItem('bar').item(0) is null
PASS item.properties.namedItem('bar')[0] is undefined.
PASS item.properties.names.length is 0
PASS item.properties.names.item(0) is null
PASS item.properties.names[0] is undefined.

Set parent.firstChild.id to 'id1'
PASS item.properties.length is 2
PASS item.properties.item(0) is parent.firstChild
PASS item.properties[0] is parent.firstChild
PASS item.properties.item(1) is parent.firstChild.firstChild
PASS item.properties[1] is parent.firstChild.firstChild
PASS item.properties.namedItem('foo').length is 1
PASS item.properties.namedItem('foo').item(0) is parent.firstChild
PASS item.properties.namedItem('foo')[0] is parent.firstChild
PASS item.properties.namedItem('bar').length is 1
PASS item.properties.namedItem('bar').item(0) is parent.firstChild.firstChild
PASS item.properties.namedItem('bar')[0] is parent.firstChild.firstChild
PASS item.properties.names.length is 2
PASS item.properties.names.item(0) is 'foo'
PASS item.properties.names[0] is 'foo'
PASS item.properties.names.item(1) is 'bar'
PASS item.properties.names[1] is 'bar'

Remove id attribute from parent.firstChild
PASS item.properties.length is 0
PASS item.properties.item(0) is null
PASS item.properties[0] is undefined.
PASS item.properties.namedItem('foo').length is 0
PASS item.properties.namedItem('foo').item(0) is null
PASS item.properties.namedItem('foo')[0] is undefined.
PASS item.properties.namedItem('bar').length is 0
PASS item.properties.namedItem('bar').item(0) is null
PASS item.properties.namedItem('bar')[0] is undefined.
PASS item.properties.names.length is 0
PASS item.properties.names.item(0) is null
PASS item.properties.names[0] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
